Feat: 카카오 로그인 변수 수정

+++ src/main/java/com/project/final_project/login/kakao/repository/KakaoLoginLogRepository.java
@@ -7,7 +7,7 @@ import org.springframework.data.jpa.repository.JpaRepository;
 import org.springframework.data.jpa.repository.Query;
 import org.springframework.data.repository.query.Param;
 
-public interface KakaoLoginLogRepository extends JpaRepository<KakaoLoginLog, Long> {
+public interface KakaoLoginLogRepository extends JpaRepository<KakaoLoginLog, Integer> {
   @Query("select k from KakaoLoginLog k where k.state = :state")
   KakaoLoginLog getKakaoLoginLogByState(@Param("state") String state);
 }
